pyproject.toml
[project]
name = "text2cypher"
version = "0.1.0"
description = "Clinical note generation with T5 and PEFT"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  # --- Core ML stack (versions known to work together) ---
  "torch>=2.2,<2.3",            # 2.2.x has device_mesh etc.
  "torchvision>=0.17,<0.18",
  "torchaudio>=2.2,<2.3",
  "transformers>=4.42,<4.43",
  "accelerate>=0.33,<0.34",
  "peft>=0.11,<0.12",
  "datasets>=2.20,<2.21",
  "tokenizers>=0.19,<0.20",
  "safetensors>=0.4,<0.5",

  # --- Your existing deps (left mostly as-is) ---
  "pytorch-lightning>=2.0,<3.0",
  "evaluate",
  "pydantic>=2",
  "langsmith",
  "langchain-openai",
  "hydra-core>=1.3",
  "loguru",
  "wandb",
  "omegaconf>=2.3",
  "numpy<2.0",
  "nltk",
  "rouge-score",
  "bert-score",
  "pandas",
  "s3fs",
  "boto3",
  "sagemaker",
  "python-dotenv",
  "fastapi",
  "uvicorn[standard]",
  "requests",
  "psycopg2-binary",
]

[project.optional-dependencies]
dev = [
  "black",
  "ruff",
  "pytest",
  "pytest-cov",
  "pytest-mock",
  "mypy",
  "ipykernel"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.ruff]
line-length = 88
target-version = "py39"
select = ["E", "F", "I"]
ignore = ["E501"]  # Ignore line too long

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q"
testpaths = [
  "tests",
]
markers = [
  "integration: marks tests as integration (deselect with '-m \"not integration\"')"
]

[tool.mypy]
python_version = "3.9"
ignore_missing_imports = true
